tst: split fixture into odb-only DbFixture and STA Fixture
The tst Fixture unconditionally created a dbSta instance, so every test
that used it linked OpenSTA + dbSta even when it only manipulated an odb
database (e.g. the odb helper-based tests). OpenSTA/dbSta debug symbols
dominate those binaries (~160 MB each).

Split Fixture into a base tst::DbFixture (tst/db_fixture.h, odb-only: db
manipulation, LEF tech loading, file lookup) and a derived tst::Fixture
(tst/fixture.h) that adds the dbSta/OpenSTA instance, readLiberty, and
the LEF library loaders that notify STA. The base lives in a new tst_base
library (//src/tst:db_fixture in Bazel) that does not link STA.

odb DB-only tests and SimpleDbFixture now derive from DbFixture and link
tst_base; STA consumers (PDK fixtures, IntegratedFixture, rsz/dbSta/est
tests) are unchanged. Mirrored in Bazel.
