+++ src/odb/test/cpp/BUILD
@@ -47,7 +47,7 @@ cc_test(
     deps = [
         "//src/odb/src/db",
         "//src/odb/test/cpp/helper",
-        "//src/tst",
+        "//src/tst:db_fixture",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
         "@spdlog",
@@ -102,7 +102,7 @@ cc_test(
         "//src/odb/src/db",
         "//src/odb/src/gdsin",
         "//src/odb/src/gdsout",
-        "//src/tst",
+        "//src/tst:db_fixture",
         "//src/utl",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
@@ -235,7 +235,7 @@ cc_test(
     srcs = ["TestObjectType.cpp"],
     deps = [
         "//src/odb/src/db",
-        "//src/tst",
+        "//src/tst:db_fixture",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
     ],
@@ -271,7 +271,7 @@ cc_test(
         "//src/odb/src/3dblox",
         "//src/odb/src/db",
         "//src/odb/test/cpp/helper",
-        "//src/tst",
+        "//src/tst:db_fixture",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
     ],
@@ -292,7 +292,7 @@ cc_test(
     deps = [
         "//src/odb/src/3dblox",
         "//src/odb/src/db",
-        "//src/tst",
+        "//src/tst:db_fixture",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
     ],
@@ -309,7 +309,7 @@ cc_test(
     deps = [
         "//src/odb/src/3dblox",
         "//src/odb/src/db",
-        "//src/tst",
+        "//src/tst:db_fixture",
         "@googletest//:gtest",
         "@googletest//:gtest_main",
     ],
