+++ src/odb/test/cpp/TestCallBacks.cpp
@@ -3,7 +3,7 @@
 #include "helper.h"
 #include "odb/db.h"
 #include "odb/dbTypes.h"
-#include "tst/fixture.h"
+#include "tst/db_fixture.h"
 
 namespace odb {
 namespace {
