+++ src/odb/test/cpp/helper/CMakeLists.txt
@@ -7,7 +7,7 @@ add_library(odb_test_helper helper.cpp)
 target_link_libraries(odb_test_helper
   PUBLIC
     db
-    tst
+    tst_base
 )
